Goodbye sbt, hello Mill.
