+++ florence-core/src/florence/core/model/styling/WithCommonProps.scala
