+++ florence-core/src/florence/core/rendering/Renderer.scala
