+++ florence-core/src/florence/package.scala
