+++ florence-core/test/src-js/florence/PlaceholderJSTests.scala
@@ -1,0 +1,11 @@
+package florence
+
+class PlaceholderJSTests extends munit.FunSuite {
+  
+  // TODO: Write a real test...
+
+  test("Core (placeholder JS)") {
+    assert(1 == 1)
+  }
+
+}
