+++ florence-core/test/src/florence/PlaceholderTests.scala
@@ -1,0 +1,11 @@
+package florence
+
+class PlaceholderTests extends munit.FunSuite {
+  
+  // TODO: Write a real test...
+
+  test("Core (placeholder)") {
+    assert(1 == 1)
+  }
+
+}
