+++ florence-renderer/test/src/florence/renderer/PlaceholderTests.scala
@@ -1,0 +1,11 @@
+package florence
+
+class PlaceholderTests extends munit.FunSuite {
+  
+  // TODO: Write a real test...
+
+  test("Renderer JS (placeholder)") {
+    assert(1 == 1)
+  }
+
+}
