--- project/build.properties
@@ -1,1 +1,0 @@
-sbt.version = 1.11.6
