--- project/plugins.sbt
@@ -1,10 +1,0 @@
-addSbtPlugin("org.scala-js"        % "sbt-scalajs"            % "1.20.1")
-addSbtPlugin("com.eed3si9n"        % "sbt-projectmatrix"      % "0.11.0")
-addSbtPlugin("org.scalameta"       % "sbt-scalafmt"           % "2.5.5")
-addSbtPlugin("org.typelevel"       % "sbt-typelevel-scalafix" % "0.8.0")
-addSbtPlugin("com.github.sbt"      % "sbt-ci-release"         % "1.11.2")
-addSbtPlugin("com.alejandrohdezma" % "sbt-github"             % "0.12.0")
-addSbtPlugin("com.alejandrohdezma" % "sbt-github-header"      % "0.12.0")
-addSbtPlugin("com.alejandrohdezma" % "sbt-github-mdoc"        % "0.12.0")
-addSbtPlugin("org.scalameta"       % "sbt-mdoc"               % "2.7.2")
-addSbtPlugin("de.heikoseeberger"   % "sbt-header"             % "5.10.0")
