+++ sandbox-tyrian/src/florence/sandbox/WeatherData.scala
