+++ sandbox-tyrian/tyrianapp.js
@@ -1,5 +1,5 @@
 import {
   TyrianApp
-} from './target/scala-3.6.4/tyrian-sandbox-fastopt/main.js';
+} from './../out/sandbox-tyrian/fastLinkJS.dest/main.js';
 
 TyrianApp.launch("myapp");
