+++ sandbox/index.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="UTF-8">
+    <title>Florence Demo</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            margin: 20px;
+        }
+        canvas {
+            border: 1px solid #ccc;
+            margin-top: 20px;
+        }
+    </style>
+</head>
+<body>
+<h1>Florence Chart Demo</h1>
+<p>This is a demo of the Florence charting library with Scala.js.</p>
+
+<canvas id="chart-canvas" width="800" height="400"></canvas>
+
+<script type="module" src="./../out/sandbox/fastLinkJS.dest/main.js"></script>
+</body>
+</html>
